daemon: move datapath-related functions to `datapath.go`
Signed-off by: Ian Vernon <ian@cilium.io>
